requirements/base.txt
# Core Django
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Cache
redis==5.0.1
django-redis==5.4.0

# Machine Learning
torch==2.1.2
torchvision==0.16.2
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

# Data Processing
python-dateutil==2.8.2

# API Documentation
drf-spectacular==0.27.0

# Utilities
python-slugify==8.0.1
Pillow==10.1.0

# Monitoring
sentry-sdk==1.39.2
